+++ common/pinot/pinotClient.go
@@ -26,12 +26,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"reflect"
-	"time"
 
 	"github.com/startreedata/pinot-client-go/pinot"
 
 	"github.com/uber/cadence/common/log"
-	"github.com/uber/cadence/common/log/tag"
 	p "github.com/uber/cadence/common/persistence"
 	"github.com/uber/cadence/common/types"
 )
@@ -93,100 +91,6 @@ func (c *PinotClient) GetTableName() string {
 }
 
 /****************************** Response Translator ******************************/
-
-func buildMap(hit []interface{}, columnNames []string) (map[string]interface{}, map[string]interface{}) {
-	systemKeyMap := make(map[string]interface{})
-	customKeyMap := make(map[string]interface{})
-
-	for i := 0; i < len(columnNames); i++ {
-		key := columnNames[i]
-		// checks if it is system key, if yes, put it into the system map; otherwise put it into custom map
-		ok, _ := isSystemKey(key)
-		if ok {
-			systemKeyMap[key] = hit[i]
-		} else {
-			customKeyMap[key] = hit[i]
-		}
-	}
-
-	return systemKeyMap, customKeyMap
-}
-
-// VisibilityRecord is a struct of doc for deserialization
-type VisibilityRecord struct {
-	WorkflowID    string
-	RunID         string
-	WorkflowType  string
-	DomainID      string
-	StartTime     int64
-	ExecutionTime int64
-	CloseTime     int64
-	CloseStatus   int
-	HistoryLength int64
-	Encoding      string
-	TaskList      string
-	IsCron        bool
-	NumClusters   int16
-	UpdateTime    int64
-	Attr          string
-}
-
-func (c *PinotClient) convertSearchResultToVisibilityRecord(hit []interface{}, columnNames []string) *p.InternalVisibilityWorkflowExecutionInfo {
-	if len(hit) != len(columnNames) {
-		return nil
-	}
-
-	systemKeyMap, customKeyMap := buildMap(hit, columnNames)
-	jsonSystemKeyMap, err := json.Marshal(systemKeyMap)
-	if err != nil { // log and skip error
-		c.logger.Error("unable to marshal systemKeyMap",
-			tag.Error(err), //tag.ESDocID(fmt.Sprintf(columnNameToValue["DocID"]))
-		)
-		return nil
-	}
-
-	var source *VisibilityRecord
-	err = json.Unmarshal(jsonSystemKeyMap, &source)
-	if err != nil { // log and skip error
-		c.logger.Error("unable to Unmarshal systemKeyMap",
-			tag.Error(err), //tag.ESDocID(fmt.Sprintf(columnNameToValue["DocID"]))
-		)
-		return nil
-	}
-
-	record := &p.InternalVisibilityWorkflowExecutionInfo{
-		DomainID:         source.DomainID,
-		WorkflowType:     source.WorkflowType,
-		WorkflowID:       source.WorkflowID,
-		RunID:            source.RunID,
-		TypeName:         source.WorkflowType,
-		StartTime:        time.UnixMilli(source.StartTime), // be careful: source.StartTime is in milisecond
-		ExecutionTime:    time.UnixMilli(source.ExecutionTime),
-		TaskList:         source.TaskList,
-		IsCron:           source.IsCron,
-		NumClusters:      source.NumClusters,
-		SearchAttributes: customKeyMap,
-	}
-	if source.UpdateTime != 0 {
-		record.UpdateTime = time.UnixMilli(source.UpdateTime)
-	}
-	if source.CloseTime != 0 {
-		record.CloseTime = time.UnixMilli(source.CloseTime)
-		record.Status = toWorkflowExecutionCloseStatus(source.CloseStatus)
-		record.HistoryLength = source.HistoryLength
-	}
-
-	return record
-}
-
-func toWorkflowExecutionCloseStatus(status int) *types.WorkflowExecutionCloseStatus {
-	if status < 0 {
-		return nil
-	}
-	closeStatus := types.WorkflowExecutionCloseStatus(status)
-	return &closeStatus
-}
-
 func (c *PinotClient) getInternalListWorkflowExecutionsResponse(
 	resp *pinot.BrokerResponse,
 	isRecordValid func(rec *p.InternalVisibilityWorkflowExecutionInfo) bool,
@@ -208,7 +112,8 @@ func (c *PinotClient) getInternalListWorkflowExecutionsResponse(
 	response.Executions = make([]*p.InternalVisibilityWorkflowExecutionInfo, 0)
 
 	for i := 0; i < numOfActualHits; i++ {
-		workflowExecutionInfo := c.convertSearchResultToVisibilityRecord(actualHits[i], columnNames)
+
+		workflowExecutionInfo := ConvertSearchResultToVisibilityRecord(actualHits[i], columnNames, c.logger, isSystemKey)
 
 		if isRecordValid == nil || isRecordValid(workflowExecutionInfo) {
 			response.Executions = append(response.Executions, workflowExecutionInfo)
@@ -249,7 +154,7 @@ func (c *PinotClient) getInternalGetClosedWorkflowExecutionResponse(resp *pinot.
 	schema := resp.ResultTable.DataSchema // get the schema to map results
 	columnNames := schema.ColumnNames
 	actualHits := resp.ResultTable.Rows
-	response.Execution = c.convertSearchResultToVisibilityRecord(actualHits[0], columnNames)
+	response.Execution = ConvertSearchResultToVisibilityRecord(actualHits[0], columnNames, c.logger, isSystemKey)
 
 	return response, nil
 }
